+++ lib/src/base/os_service_none.dart
@@ -1,0 +1,8 @@
+import 'package:dart_bones/dart_bones.dart';
+
+import 'base_logger.dart';
+
+/// Holds the information about the current user.
+class OsService {
+  OsService(BaseLogger logger);
+}