+++ test/base/crypto_engine_test.dart
@@ -1,5 +1,5 @@
 import 'package:dart_bones/dart_bones.dart';
-import 'package:dart_bones/src/base/crypto_engine.dart';
+import 'package:dart_bones/src/base/crypto_engine_io.dart';
 import 'package:dart_bones/src/base/kiss_random.dart';
 import 'package:test/test.dart';